bundler: inline ungate_support into bundle_v2 and lib, delete the file
ungate_support.rs was a porting-reference grab-bag holding value types
extracted from bundle_v2.zig plus crate-name alias shims, glob-re-exported
at the crate root.

Move the value types (PartRange, StableRef, ImportTracker, CompileResult,
CompileResultForSourceMap, ContentHasher, WrapKind, DeclInfo, CrossChunkImport,
the entry_point and js_meta modules, generic_path_with_pretty_initialized,
target_from_hashbang) into bundle_v2.rs next to their bundle_v2.zig
counterparts. Move the crate-name aliases (bun_fs, bun_css, perf, bun_renamer,
html_import_manifest, html_scanner, cross_chunk_import, index, part,
import_record, JSAst, UseDirective, MangledProps, PathTemplate, EventLoop)
to the crate root in lib.rs. Drop the unused bun_str and bun_output aliases.

Delete ungate_support.rs and strip the porting-reference comments. Item
visibilities and the crate-root spellings callers use are unchanged.
