+++ src/bundler/linker_context/generateCodeForFileInChunkJS.rs
@@ -7,9 +7,9 @@ use bun_collections::{BoundedArray, VecExt};
 use bun_js_printer::renamer;
 use bun_js_printer::{self as js_printer, PrintResult, PrintResultSuccess};
 
+use crate::generic_path_with_pretty_initialized;
 use crate::linker_context_mod::{StmtList, StmtListWhich};
 use crate::options::Format as OutputFormat;
-use crate::ungate_support::generic_path_with_pretty_initialized;
 use crate::{Chunk, DeclInfo, DeclInfoKind, Index, LinkerContext, Part, PartRange, WrapKind};
 
 use bun_ast::StoreRef;
